# Optimization: statements reuse previous column name